+++ src/main/index.ts
@@ -56,6 +56,7 @@ import { TrackpadGestureLedger } from './trackpad-gesture'
 import { registerConverterIpc } from '../core/converter/register-ipc'
 import { registerNodeDependencyIpc } from '../core/node-dependencies/register-ipc'
 import { registerOllamaIpc } from '../core/ollama/register-ipc'
+import { registerOpenWebUiHosting } from './open-webui-hosting'
 import { registerMinecraftIpc } from '../core/minecraft/register-ipc'
 import { registerAwsIdentityIpc } from '../core/aws-identity'
 import { registerAwsResourceIpc } from '../core/aws-resource-register-ipc'
@@ -2378,6 +2379,7 @@ app.whenReady().then(async () => {
   registerConverterIpc(corePlatform)
   const nodeDependencyService = registerNodeDependencyIpc(corePlatform)
   registerOllamaIpc(corePlatform)
+  registerOpenWebUiHosting(getMainWindow, app.getPath('userData'))
   minecraftServers = registerMinecraftIpc(corePlatform).manager
   registerAwsIdentityIpc(corePlatform, {
     resolveAwsCli: async () => {